requirements.txt
openai
numpy
pandas
PyPDF2
sentence-transformers
python-dotenv
sqlite3